+++ CHANGELOG.md
@@ -2,9 +2,9 @@ HumHub Changelog
 ================
 
 1.18.0-beta.5 (Unreleased)
---------------------------
+----------------------------------
+- Enh #7719: Make Add/Remove Group Member optional for group managers
 - Fix #7720: Saving the new Mobile App admin form when no whitelisted domain
-
 
 1.18.0-beta.4 (September 19, 2025)
 ----------------------------------